Integrate delivery bar chart with V&L section
